move to util folder
